+++ frontends/web/component/react/index.ts
@@ -1,6 +1,7 @@
 export * from "./boytacean";
 export * from "./boytacean-context";
 export * from "./boytacean-gamepad";
+export * from "./boytacean-hooks";
 export * from "./boytacean-keyboard";
 export * from "./boytacean-provider";
 export * from "./boytacean-screen";